# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 隐私邮箱外部取件 API v2
  version: 2.0.0
  description: |
    本文档主要描述外部接口，并记录升级所需的少量管理兼容接口。服务提供两种
    取件能力：标准只读 IMAPS，以及使用 API Key 或派生取码 URL 读取验证码历史。

    管理端可为每个主号配置上游隐式 TLS IMAP 主机和端口；创建时默认使用
    `imap.mail.me.com:993`。修改规范化后的主机、端口或用户名代表切换邮箱来源，会清理
    旧来源的同步游标、兼容快照、消费历史、待处理 `\\Seen` 任务、v2 归档和 OTP 历史，
    轮换公开 IMAPS 的 `UIDVALIDITY`，并将同步状态重置为 `pending`。

    主号也可以使用 `custom` 邮箱模式并保存邮箱后缀。自定义随机邮箱使用 8–12 位
    ASCII 小写英文字母和数字作为本地部分，批量内和数据库中均禁止重复；该流程不调用 Apple，原 iCloud
    自动创建计划和限流规则保持不变。

    收件路由有三套：`mailbox_type=icloud` 加默认 `imap.mail.me.com:993` 是直连 iCloud IMAP；
    `mailbox_type=icloud` 加非默认 IMAP 来源（主机/端口或用户名不同于 iCloud 默认）是 iCloud 隐私邮箱转发到第三方 IMAP；
    `mailbox_type=custom` 是自定义域名邮箱，按原始投递收件人头匹配本地地址。

    如果 iCloud 隐私邮箱转发到第三方邮箱，仍应保持 `mailbox_type=icloud`，并将上游
    `imap_username` 填为第三方 IMAP 登录/最终物理投递目标地址（若该邮箱服务允许使用邮箱地址登录）。同步会优先使用
    `X-ICLOUD-HME` 的 `p=`/`f=` 路由信息，再结合 `Delivered-To`、`X-Original-To`、
    `Original-Recipient` 和常见 Apple 原始收件人头恢复隐私邮箱归属；Apple 的 `f=` 可以是中间
    转发地址，不必与最终 `Delivered-To` 相同。若只在 `To/Cc` 留下隐私地址，必须显式启用弱收件人头才会匹配。
    这些规则只控制邮件取件与归属；Apple 目录同步和自动创建仍使用现有的主号默认转发目标校验。

    IMAPS 用户名是隐私邮箱地址。客户端可以使用本服务签发的 IMAP 密码登录，
    或通过 `POST /oauth2/v2.0/token` 获取一小时访问令牌后使用 XOAUTH2。
    邮箱只有只读 INBOX，支持 LIST、STATUS、SELECT/EXAMINE、SEARCH、
    FETCH/UID FETCH 和 IDLE；所有写操作返回只读错误。

    新 v2 alias 的 API Key、IMAP 密码、client ID、refresh token 与 access token
    均由本服务签发。legacy alias 的兼容轮换只替换 API Key 和由其派生的旧直达链接，
...
    管理员可显式执行全量凭据轮换：所有 legacy alias 会强制迁移为 v2，所有现有
...
servers:
...
paths:
  /api/v1/pool/leases/{leaseID}/code:
    get:
      tags: [Mailbox Pool]
      operationId: readPoolLeaseCode
      summary: 读取项目所领取邮箱的最新验证码，可选有界等待
      description: |
        Pool Key 必须属于 lease 的项目。lease 为有效 leased 或 used，且主号和 alias 启用。
        查询起点取 after 与领取时间较晚者，只返回严格晚于起点的验证码。
        wait_seconds 默认 0，保留原单次按需读取；启用共享收件时正值允许等待新归档，
        预算包含首轮同步。正常到期仍返回 no_code，首轮同步未完成或上游错误返回 503。
        未配置共享通知时退回单次读取。等待不持有全局轮换/Pool 锁，最终重新校验归属及状态。
        共享模式同 root 容量 2，每 2 秒恢复 1 次额度；最多一个普通读取和一个长等待，
        等待期间额度恢复，完成后不追加冷却。旧非共享路径保持独占和完成后 2 秒冷却。
        超额返回 429；被拦截请求不延长恢复时间。
      security:
        - poolKey: []
      parameters:
        - name: leaseID
          in: path
          required: true
          schema: {type: string}
        - name: after
          in: query
          schema: {type: string, format: date-time}
        - name: wait_seconds
          in: query
          description: 单个十进制整数；默认 0。HTTP 超时应留网络收尾余量。
          schema: {type: integer, minimum: 0, maximum: 15, default: 0}
      responses:
        "200":
          description: 命中新验证码或本次等待范围内无新码；不返回邮箱凭据。
          headers:
            Cache-Control:
              schema: {type: string, const: no-store}
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    oneOf:
                      - type: object
                        required: [success, otp, time, email]
                        properties:
                          success: {type: boolean, const: true}
                          otp: {type: string}
                          time: {type: string, format: date-time}
                          email: {type: string, format: email}
                      - type: object
                        required: [success, code, retryable, email]
                        properties:
                          success: {type: boolean, const: false}
                          code: {type: string, const: no_code}
                          retryable: {type: boolean, const: true}
                          email: {type: string, format: email}
        "400":
          description: INVALID_POOL_REQUEST，参数非法或 wait_seconds 重复。
        "401":
          description: INVALID_POOL_KEY，项目 Key 无效、已轮换或停用。
        "404":
          description: NOT_FOUND，lease 不存在或属于其他项目。
        "409":
          description: LEASE_CLOSED，领取关闭、过期或主号/alias 停用。
        "429":
          description: RATE_LIMITED，同类取件在途、速率预算不足或容量已满。在途/容量原因返回建议复查间隔。
          headers:
            Retry-After:
              description: 剩余等待时间向上取整的秒数。
              schema: {type: string, pattern: '^[1-9][0-9]*$'}
            X-Retry-After-Ms:
              description: 剩余等待时间或建议复查间隔，单位毫秒。
              schema: {type: string, pattern: '^[1-9][0-9]*$'}
            X-RateLimit-Reason:
              schema:
                type: string
                enum: [mailbox_rate, mailbox_waiting, mailbox_inflight, mailbox_cooldown, account_concurrency, global_concurrency, global_rate, mailbox_capacity]
        "503":
          description: SYNC_UNAVAILABLE 或 POOL_UNAVAILABLE；同步异常时 Retry-After 为 3。
...
  /api/v1/accounts:
    servers:
      - url: https://HOST/{admin_path}
        description: 首选的安装级随机管理 API 前缀；将 admin_path 替换为 keys 卷中值去掉首尾斜杠后的形式
        variables:
          admin_path:
            default: 00000000000000000000000000000000/admin
            description: 首次启动生成并持久化的管理路径，替换时去掉文件值的开头和结尾斜杠
      - url: https://HOST/admin
        description: 升级前客户端使用的固定兼容入口；认证、CSRF 和限流规则相同
    get:
      tags:
        - Admin Accounts
      operationId: listAdminAccounts
      summary: 查询主号列表
      description: |
        按主号邮箱、ID 升序返回分页列表。`query` 会先去除首尾空白，再对主号邮箱、自定义
        邮箱后缀和备注执行不区分大小写的字面量子串匹配；空查询返回全部主号。响应中的
        `pagination.total` 是当前查询条件下的主号总数，`has_more` 表示当前页之后是否还有结果。
      security:
        - adminSession: []
      parameters:
        - name: query
          in: query
          required: false
          description: 主号邮箱、自定义邮箱后缀或备注的搜索词；服务会先去除首尾空白，处理后的值最多 200 个字符。
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: 本页最多返回的主号数。
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 20
        - name: offset
          in: query
          required: false
          description: 在当前查询结果中跳过的主号数。
          schema:
            type: integer
            minimum: 0
            maximum: 1000000
            default: 0
      responses:
        "200":
          description: 当前查询条件下的一页主号
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/AdminNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountListResponse"
        "400":
          description: query、limit 或 offset 参数无效
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/AdminNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAppleErrorResponse"
              examples:
                invalidLimit:
                  value:
                    error:
                      code: VALIDATION_FAILED
                      message: limit 参数无效
                      request_id: 6oe7j0x3x3pJv9Dv
                queryTooLong:
                  value:
                    error:
                      code: VALIDATION_FAILED
                      message: query 参数不能超过 200 个字符
                      request_id: 6oe7j0x3x3pJv9Dv
        "401":
          $ref: "#/components/responses/AdminAuthUnauthorized"
        "500":
          $ref: "#/components/responses/AdminInternalError"
    post:
      tags:
        - Admin Accounts
      operationId: createAdminAccount
      summary: 添加主号
      description: |
        添加用于主号级邮件同步的账户。`imap_host` 和 `imap_port` 可分别指定隐式 TLS
        IMAP 服务；省略时分别使用 `imap.mail.me.com` 和 `993`，显式传入 `null` 则返回 400。
        iCloud 模式保持原有 Apple 规则；若邮件由 iCloud 转发到第三方邮箱，仍选择 `icloud`，并应
        将 `imap_username` 填为第三方 IMAP 登录/最终物理投递目标地址，以便校验投递标头。选择 `custom` 后须填写 `email_suffix`，
        并在主号详情中按数量生成 8–12 位小写英文字母和数字随机邮箱。IMAP 密码只会加密保存，不会出现在响应或日志中。
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/AdminCSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateAccountRequest"
            example:
              name: 个人主号
              email: owner@icloud.com
              mailbox_type: icloud
              imap_host: imap.mail.me.com
              imap_port: 993
              imap_username: owner@icloud.com
              imap_password: APP_SPECIFIC_PASSWORD
      responses:
        "201":
          description: 主号创建成功
          headers:
            Location:
              description: 新主号的管理 API 相对路径。
              schema:
                type: string
                format: uri-reference
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/AdminNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountResponse"
        "400":
          $ref: "#/components/responses/AdminAccountBadRequest"
        "401":
          $ref: "#/components/responses/AdminAuthUnauthorized"
        "403":
          $ref: "#/components/responses/AdminCSRFForbidden"
        "409":
          $ref: "#/components/responses/AdminAccountConflict"
        "413":
          $ref: "#/components/responses/AdminJSONTooLarge"
        "415":
          $ref: "#/components/responses/AdminUnsupportedMediaType"
        "500":
          $ref: "#/components/responses/AdminInternalError"
  /api/v1/accounts/{id}:
    servers:
      - url: https://HOST/{admin_path}
        description: 首选的安装级随机管理 API 前缀；将 admin_path 替换为 keys 卷中值去掉首尾斜杠后的形式
        variables:
          admin_path:
            default: 00000000000000000000000000000000/admin
            description: 首次启动生成并持久化的管理路径，替换时去掉文件值的开头和结尾斜杠
      - url: https://HOST/admin
        description: 升级前客户端使用的固定兼容入口；认证、CSRF 和限流规则相同
    get:
      tags:
        - Admin Accounts
      operationId: getAdminAccount
      summary: 获取主号详情
      description: |
        返回主号、所属隐私邮箱、Apple 会话状态和自动创建计划。`aliases` 只包含当前数据库
        分页；`pagination.total` 是该主号的隐私邮箱总数，`has_more` 表示当前页之后是否还有结果。
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/AdminAccountID"
        - name: limit
          in: query
          required: false
          description: 本页最多返回的该主号隐私邮箱数。
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 20
        - name: offset
          in: query
          required: false
          description: 在该主号的隐私邮箱列表中跳过的邮箱数。
          schema:
            type: integer
            minimum: 0
            maximum: 1000000
            default: 0
      responses:
        "200":
          description: 主号详情及所属隐私邮箱的当前数据库分页
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/AdminNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountDetailResponse"
        "400":
          description: limit 或 offset 参数无效
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/AdminNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAppleErrorResponse"
              example:
                error:
                  code: VALIDATION_FAILED
                  message: limit 参数无效
                  request_id: 6oe7j0x3x3pJv9Dv
        "401":
          $ref: "#/components/responses/AdminAuthUnauthorized"
        "404":
          $ref: "#/components/responses/AdminAccountNotFound"
        "500":
          $ref: "#/components/responses/AdminInternalError"
    put:
      tags:
        - Admin Accounts
      operationId: updateAdminAccount
      summary: 更新主号
      description: |
        更新主号配置。省略 `imap_host` 或 `imap_port` 时分别保留当前值，显式传入 `null` 则
        返回 400。规范化后的 IMAP 主机、端口或用户名实际发生变化后，服务会清除该主号旧邮箱
        来源的同步游标、兼容快照、消费历史、待处理 `\Seen` 任务、v2 归档和 OTP 历史，轮换
        公开 IMAPS 的 `UIDVALIDITY`，并在下次同步确认前把主号及所属隐私邮箱的同步状态重置为
        `pending`。修改 IMAP 密码（直连 iCloud 为 App 专用密码，第三方转发为第三方 IMAP 密码）也会重置同步状态，但保留已有邮件；`imap_password` 省略、
        传 `null` 或传空字符串时均保留当前密码。
        已有隐私邮箱后，仅主号邮箱及邮箱模式/后缀不可修改；IMAP 主机、端口和用户名仍可修改。
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/AdminAccountID"
        - $ref: "#/components/parameters/AdminCSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateAccountRequest"
            example:
              name: 个人主号
              email: owner@icloud.com
              mailbox_type: icloud
              imap_host: imap.example.com
              imap_port: 993
              imap_username: owner@icloud.com
              imap_password: ""
              enabled: true
      responses:
        "200":
          description: 主号配置已更新
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/AdminNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountResponse"
        "400":
          $ref: "#/components/responses/AdminAccountBadRequest"
        "401":
          $ref: "#/components/responses/AdminAuthUnauthorized"
        "403":
          $ref: "#/components/responses/AdminCSRFForbidden"
        "404":
          $ref: "#/components/responses/AdminAccountNotFound"
        "409":
          $ref: "#/components/responses/AdminAccountConflict"
        "413":
          $ref: "#/components/responses/AdminJSONTooLarge"
        "415":
          $ref: "#/components/responses/AdminUnsupportedMediaType"
        "500":
          $ref: "#/components/responses/AdminInternalError"
    delete:
      tags:
        - Admin Accounts
      operationId: deleteAdminAccount
      summary: 删除主号
      description: |
        永久删除本地主号及其所属隐私邮箱、API Key、邮件快照、消费历史、待处理 `\Seen` 任务、
        v2 归档元数据和 MIME 文件、IMAP 同步游标、Apple Web 会话和自动创建计划。该操作只
        清理本地数据，不会删除 Apple Hide My Email 中的远端地址。
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/AdminAccountID"
        - $ref: "#/components/parameters/AdminCSRFToken"
      responses:
        "204":
          description: 本地主号及其关联数据已删除
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/AdminNoStore"
        "401":
          $ref: "#/components/responses/AdminAuthUnauthorized"
        "403":
          $ref: "#/components/responses/AdminCSRFForbidden"
        "404":
          $ref: "#/components/responses/AdminAccountNotFound"
        "500":
          $ref: "#/components/responses/AdminInternalError"
...
  /oauth2/v2.0/token:
    post:
      summary: 用 refresh token 换取 IMAPS XOAUTH2 访问令牌
      description: |
        此接口是只读 IMAPS 的配套认证接口，不构成第三种取件方式。
        client ID 与 refresh token 都属于同一个隐私邮箱并由本服务签发。
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: false
              required: [grant_type, client_id, refresh_token]
              properties:
                grant_type:
                  type: string
                  const: refresh_token
                client_id:
                  type: string
                refresh_token:
                  type: string
      responses:
        "200":
          description: 一小时有效的 Bearer access token。
          headers:
            Cache-Control:
              schema:
                type: string
              example: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
              example:
                access_token: ACCESS_TOKEN
                token_type: Bearer
                expires_in: 3600
        "400":
          $ref: "#/components/responses/OAuthInvalidRequest"
        "401":
          $ref: "#/components/responses/OAuthInvalidClient"
        "429":
          $ref: "#/components/responses/OAuthUnavailable"
        "503":
          $ref: "#/components/responses/OAuthUnavailable"
...
components:
  securitySchemes:
    externalOAuthToken:
      type: http
      scheme: bearer
      bearerFormat: deployment OAuth token
    aliasAPIKey:
      type: http
      scheme: bearer
      bearerFormat: service-issued API Key
    derivedToken:
      type: apiKey
      in: query
      name: token
      description: 仅授权只读 OTP 历史；不能用于消费 recent-mail。
    aliasAPIKeyQuery:
      type: apiKey
      in: query
      name: api_key
    poolKey:
      type: http
      scheme: bearer
      bearerFormat: pool project Key
      description: 以 `pool_` 开头的项目 Key；仅与管理员 Session 组合时可调用 Pool 已使用邮箱退休删除接口。
    adminSession:
      type: apiKey
      in: cookie
      name: icloud_admin_session
...
  headers:
    RequestID:
      description: 服务端生成的请求追踪标识；错误响应体中的 request_id 与此值一致。
      schema:
        type: string
      example: 6oe7j0x3x3pJv9Dv
    AdminNoStore:
      description: 禁止保存包含后台状态或一次性凭据的响应。
      schema:
        type: string
        const: no-store, private
      example: no-store, private
...
  schemas:
    AdminAccountListResponse:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/AdminAccountPage"
    AdminAppleErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/AdminAppleError"
...
  responses:
    AdminAuthUnauthorized:
      description: 后台会话 Cookie 缺失、无效或已经过期
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/AdminNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AdminAppleErrorResponse"
          examples:
            loginRequired:
              value:
                error:
                  code: AUTH_REQUIRED
                  message: 请先登录
                  request_id: 6oe7j0x3x3pJv9Dv
            sessionExpired:
              value:
                error:
                  code: SESSION_EXPIRED
                  message: 登录会话已过期
                  request_id: 6oe7j0x3x3pJv9Dv
...